refactor: split state classes into individual files
