+++ acq4/devices/PatchPipette/states/out.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+from ._base import PatchPipetteState
+
+
+class OutState(PatchPipetteState):
+    stateName = 'out'
+
+    _parameterDefaultOverrides = {
+        'initialPressureSource': 'atmosphere',
+        'initialClampMode': 'VC',
+        'initialVCHolding': 0,
+        'initialTestPulseEnable': False,
+        'finishPatchRecord': True,
+    }
