+++ acq4/devices/PatchPipette/states/whole_cell.py
@@ -1,0 +1,32 @@
+from __future__ import annotations
+
+from acq4.util import ptime
+from ._base import PatchPipetteState
+
+
+class WholeCellState(PatchPipetteState):
+    stateName = 'whole cell'
+    _parameterDefaultOverrides = {
+        'initialPressureSource': 'atmosphere',
+        'initialClampMode': 'VC',
+        'initialVCHolding': -70e-3,
+        'initialTestPulseEnable': True,
+        'initialAutoBiasEnable': True,
+        'initialAutoBiasTarget': -70e-3,
+    }
+
+    def run(self):
+        patchrec = self.dev.patchRecord()
+        patchrec['wholeCellStartTime'] = ptime.time()
+        patchrec['wholeCellPosition'] = tuple(self.dev.pipetteDevice.globalPosition())
+
+        # TODO: Option to switch to I=0 for a few seconds to get initial RMP decay
+
+        while True:
+            # TODO: monitor for cell loss
+            self.sleep(0.1)
+
+    def cleanup(self):
+        patchrec = self.dev.patchRecord()
+        patchrec['wholeCellStopTime'] = ptime.time()
+        super().cleanup()
